Revert "Convert jint to jlong for native pointers in GLFunctionTable to support 64 bit."
Landed on wrong branch.

This reverts commit 5ffecf060de7ee5e94a73dff21ebb2705d947fd9.

+++ chromium/plat_support/graphic_buffer_impl.h
@@ -30,12 +30,12 @@ class GraphicBufferImpl {
  public:
   ~GraphicBufferImpl();
 
-  static long Create(int w, int h);
-  static void Release(long buffer_id);
-  static int MapStatic(long buffer_id, AwMapMode mode, void** vaddr);
-  static int UnmapStatic(long buffer_id);
-  static void* GetNativeBufferStatic(long buffer_id);
-  static uint32_t GetStrideStatic(long buffer_id);
+  static int Create(int w, int h);
+  static void Release(int buffer_id);
+  static int MapStatic(int buffer_id, AwMapMode mode, void** vaddr);
+  static int UnmapStatic(int buffer_id);
+  static void* GetNativeBufferStatic(int buffer_id);
+  static uint32_t GetStrideStatic(int buffer_id);
 
  private:
   status_t Map(AwMapMode mode, void** vaddr);
